requirements.txt
numpy>=1.20.0
scipy>=1.7.0
soundfile>=0.10.0
Pillow>=8.0.0
hypothesis>=6.0.0
